+++ crates/blockstore/src/tests/high_level/block_store_adapter.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use async_trait::async_trait;
+use byte_unit::Byte;
 use futures::stream::BoxStream;
 use std::fmt::{self, Debug};
 
@@ -49,14 +50,14 @@ impl<B: BlockStore + Send + Sync + Debug + 'static> BlockStoreReader for BlockSt
         self.0.num_blocks().await
     }
 
-    fn estimate_num_free_bytes(&self) -> Result<u64> {
+    fn estimate_num_free_bytes(&self) -> Result<Byte> {
         self.0.estimate_num_free_bytes()
     }
 
     fn block_size_from_physical_block_size(
         &self,
-        block_size: u64,
-    ) -> Result<u64, InvalidBlockSizeError> {
+        block_size: Byte,
+    ) -> Result<Byte, InvalidBlockSizeError> {
         self.0.block_size_from_physical_block_size(block_size)
     }
